Move settings.js to src/settings.js
